+++ flake.nix
@@ -24,10 +24,6 @@
       url = "github:openclaw/nix-openclaw";
       inputs.nixpkgs.follows = "nixpkgs";
     };
-    nixos-generators = {
-      url = "github:nix-community/nixos-generators";
-      inputs.nixpkgs.follows = "nixpkgs";
-    };
     nixpkgs = {
       url = "github:NixOS/nixpkgs/nixos-unstable";
     };
@@ -72,7 +68,6 @@
       });
       packages = import ./cloud/configuration.nix {
         inherit nixpkgs self;
-        inherit (inputs) nixos-generators;
       };
       darwinConfigurations = {
         edenzim-ltmbn8v.internal.salesforce.com = inputs.nix-darwin.lib.darwinSystem {